0.0.2 Type Adjust & Rust 2024

+++ src/v2/model/users/structs/daily_challenge_user_stats.rs
@@ -4,18 +4,18 @@ use serde::{Serialize, Deserialize};
 
 #[derive(Debug, Clone, Serialize, Deserialize)]
 pub struct DailyChallengeUserStats {
-    pub daily_streak_best: u64,
-    pub daily_streak_current: u64,
+    pub daily_streak_best: u32,
+    pub daily_streak_current: u32,
     
     pub last_update: String,
     
     pub last_weekly_streak: String,
-    pub playcount: u64,
-    pub top_10p_placements: u64,
-    pub top_50p_placements: u64,
-    pub user_id: u64,
-    pub weekly_streak_best: u64,
-    pub weekly_streak_current: u64,
+    pub playcount: u32,
+    pub top_10p_placements: u32,
+    pub top_50p_placements: u32,
+    pub user_id: u32,
+    pub weekly_streak_best: u32,
+    pub weekly_streak_current: u32,
 }
 
 impl Default for DailyChallengeUserStats {
