+++ src/v2/model/users/structs/rank_history.rs
@@ -4,7 +4,7 @@ use serde::{Serialize, Deserialize};
 #[derive(Debug, Clone, Serialize, Deserialize)]
 pub struct RankHistory {
     pub mode: String,
-    pub data: Vec<u64>,
+    pub data: Vec<u32>,
 }
 
 impl Default for RankHistory {
